% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiscore-package.R
\docType{package}
\name{hiscore-package}
\alias{hiscore}
\alias{hiscore-package}
\title{hiscore: Play Games with Saved High Scores}
\description{
Play simple interactive games in the console and have your high scores saved to your local machine.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
