% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_coinflip}
\alias{play_coinflip}
\title{Play a Coinflip Game}
\usage{
play_coinflip()
}
\value{
Nothing.
}
\description{
A simple interactive game of chance in the console. The user guesses heads or
tails by typing 'H' or 'T'. The best streak is saved as a high score, which
is saved to the user's computer and is persistent between sessions.
}
\examples{
\dontrun{play_coinflip()}
}
