% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{play_coinflip}
\alias{play_coinflip}
\title{Play a Coinflip Game}
\usage{
play_coinflip()
}
\value{
Nothing.
}
\description{
An interactive game in the console. The user guesses heads or tials by typing
'H' or 'T'. The maximum run is saved as a high score. THe user can also
choose to save their game and return later.
}
\examples{
\dontrun{play_coinflip()}
}
